requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lxml>=4.9.0
html5lib>=1.1
pytz>=2023.3

# For enhanced performance
aiohttp>=3.8.0
httpx>=0.24.0

# For proxy and anti-detection support
aiohttp-proxy>=0.1.2
fake-useragent>=1.4.0
cloudscraper>=1.2.71
brotli>=1.1.0

# For advanced features (optional)
selenium>=4.11.0
playwright>=1.37.0

# Lean deployment essentials
scraperapi-sdk>=1.5.3
stripe>=12.2.0

# Production deployment
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
redis>=4.6.0
celery>=5.3.0

# Production monitoring
sentry-sdk>=1.39.0
psutil>=5.9.0
prometheus-client>=0.19.0

# Security
python-dotenv>=1.0.0
cryptography>=41.0.0